Take the terminal type as a parameter
Although this makes the typical simple use case slightly more tedious, this
provides an easy way to configure the terminal in contexts that shouldn't make
assumptions about the system environment.  For example, the upcoming test suite
will use this to force ANSI input/output from/to files, regardless of the type
of terminal the test is running in.

It also conveniently avoids a call to getenv() in the library, since that isn't
necessarily thread-safe.