Factor out VT100 escape prefix
The "universality" of hard baked string literals is nice, but I find this
escape in particular a real struggle to read (hex character escapes in C in
general are ridiculous).  So, factor it into a macro, which removes the noise
and makes it easier to see which sequences are being used.