Remove use of strcasecmp
Not that it matters much here, but this is also more efficient since the
built-in constants don't require case conversion.  It's fine to be extremely
crude with the comparison here since the constants are known to be simple
lowercase ASCII text, and the worst-case scenario with strange TERM values is
not matching, which is correct.